Modified the outline and added a Gantt chart
I updated the outline a bit. I never like that most proposals sort of
peter out, ending with an anemic discussion of personnel and work plan.
To avoid this sad state, I have moved the Impact and Relevance to the
end of the Science/Technical/Management section so that the proposal
ends on a strong argument.

I also added an example of a LaTeX based Gantt chart. This is useful for
showing schedules.

	modified:   NASA_ROSES_LaTeX_Template.tex